Conflicts:
	src/condor_collector.V6/collector_engine.cpp
	src/condor_io/condor_ipverify.cpp
	src/condor_io/sock.cpp
	src/condor_schedd.V6/schedd.cpp
	src/condor_startd.V6/startd_main.cpp
	src/condor_syscall_lib/special_stubs.cpp
	src/condor_utils/cbstp_url.cpp
	src/condor_utils/ftp_url.cpp
	src/condor_utils/http_url.cpp
	src/condor_utils/my_hostname.h

+++ src/condor_startd.V6/Starter.cpp
@@ -1073,7 +1073,7 @@ Starter::cleanupAfterGlexec()
 {
 	// remove the copy of the user proxy that we own
 	// (the starter should remove the one glexec created for it)
-	if( s_claim->client()->proxyFile() != NULL ) {
+	if( s_claim && s_claim->client()->proxyFile() != NULL ) {
 		if ( unlink( s_claim->client()->proxyFile() ) == -1) {
 			dprintf( D_ALWAYS,
 			         "error removing temporary proxy %s: %s (%d)\n",
